requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
coverage==5.5
dj-database-url==0.5.0
Django==3.2.5
django-crispy-forms==1.12.0
djangorestframework==3.12.4
gunicorn==20.1.0
httpie==2.4.0
idna==3.2
psycopg2==2.9.9
Pygments==2.9.0
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
